requirements.txt
streamlit
crewai[tools]
crewai
langchain
langchain_ollama
langchain-community
langchain-openai
langchain-huggingface
crewai_tools
duckduckgo-search
python-dotenv
pysqlite3-binary
sentence-transformers
faiss-cpu

